backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
bidict==0.23.1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
dnspython==2.7.0
docx2txt==0.8
dotenv==0.9.9
eventlet==0.39.1
Flask==2.0.3
Flask-SocketIO==5.3.1
gevent==24.11.1
gevent-websocket==0.10.1
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.163.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.5.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.1
greenlet==3.1.1
grpcio==1.71.0rc2
grpcio-status==1.71.0rc2
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==24.2
proto-plus==1.26.0
protobuf==5.29.3
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyNaCl==1.5.0
pyparsing==3.2.1
PySocks==1.7.1
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
requests==2.32.3
rsa==4.9
setuptools==76.0.0
sh==2.2.2
simple-websocket==1.1.0
sniffio==1.3.1
sqlparse==0.5.3
stem==1.8.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.3.0
websockets==14.2
Werkzeug==2.0.3
whitenoise==6.9.0
wsproto==1.2.0
zope.event==5.0
zope.interface==7.2
